:::handling exceptions

+++ db/db_user.py
@@ -1,8 +1,8 @@
-import email
 from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from schemas import UserBase
 from db.models import DbUser
+from fastapi import HTTPException, status
 
 
 def create_user(request: UserBase, db: Session):
@@ -24,11 +24,18 @@ def get_all_users(db: Session):
 
 
 def find_user_by_id(id: int, db: Session):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+    return user
 
 
 def update_user(id: int, request: UserBase, db: Session):
-    user = db.query(DbUser).filter(DbUser.id == id)
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -40,6 +47,9 @@ def update_user(id: int, request: UserBase, db: Session):
 
 def delete_user(id: int, db: Session):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     db.delete(user)
     db.commit()
     return 'ok'
